Update `Node` and `Profile`
Change all `Inbound` to `Profile`.

Handlers:

- Refine error types and messages.
- Refine logic: when processing deleting operations, check if binding
  data still exist (profiles under a node or users under a profile).
- Fix the bug that prevents binding the query parameters when clients do
  multiple queries for nodes or profiles because of the missing of tags
  and unexported fields.
- Add `nodeId` query parameter to the opration of doing multiple queries
  for profiles.
- Update the logic of `Inbound.Tag` generation.
- Return errors when association data do not exist.

Models:

- Manually combine the `Inbound` `struct` to solve the bug that
  `interface` is encoded as the BSON binary type and it cannot be
  decoded back.

APIs:

- Change the parameters of some API functions.
- Dial gRPC with TLS encryption.
- Do non-blocking dailling otherwise it returns a connection error for
  unknown reasons.

+++ handlers/node.go
@@ -1,9 +1,11 @@
 package handlers
 
 import (
-    "github.com/bionicosmos/submgr/models"
-    "github.com/gofiber/fiber/v2"
-    "go.mongodb.org/mongo-driver/mongo"
+	"github.com/bionicosmos/submgr/models"
+	"github.com/gofiber/fiber/v2"
+	"go.mongodb.org/mongo-driver/bson"
+	"go.mongodb.org/mongo-driver/bson/primitive"
+	"go.mongodb.org/mongo-driver/mongo"
 )
 
 func FindNode(c *fiber.Ctx) error {
@@ -19,14 +21,14 @@ func FindNode(c *fiber.Ctx) error {
 
 func FindNodes(c *fiber.Ctx) error {
     var query struct {
-        skip  int64
-        limit int64
+        Skip  int64 `query:"skip"`
+        Limit int64 `query:"limit"`
     }
     if err := c.QueryParser(&query); err != nil {
-        return err
+        return fiber.NewError(fiber.StatusBadRequest, err.Error())
     }
 
-    nodes, err := models.FindNodes(query.skip, query.limit)
+    nodes, err := models.FindNodes(query.Skip, query.Limit)
     if err != nil {
         return err
     }
@@ -39,7 +41,7 @@ func FindNodes(c *fiber.Ctx) error {
 func InsertNode(c *fiber.Ctx) error {
     var node models.Node
     if err := c.BodyParser(&node); err != nil {
-        return err
+        return fiber.NewError(fiber.StatusBadRequest, err.Error())
     }
     if err := node.Insert(); err != nil {
         return err
@@ -50,7 +52,7 @@ func InsertNode(c *fiber.Ctx) error {
 func UpdateNode(c *fiber.Ctx) error {
     var node models.Node
     if err := c.BodyParser(&node); err != nil {
-        return err
+        return fiber.NewError(fiber.StatusBadRequest, err.Error())
     }
     if err := node.Update(c.Params("id")); err != nil {
         return err
@@ -59,6 +61,23 @@ func UpdateNode(c *fiber.Ctx) error {
 }
 
 func DeleteNode(c *fiber.Ctx) error {
+    id := c.Params("id")
+    nodeId, err := primitive.ObjectIDFromHex(id)
+    if err != nil {
+        return fiber.NewError(fiber.StatusBadRequest, err.Error())
+    }
+    // Find profiles whose nodeId equals to the id from parameters.
+    // If there are errors, return them.
+    // If the profile is not empty, also return an error.
+    profiles, err := models.FindProfiles(bson.D{
+        {Key: "nodeId", Value: nodeId},
+    }, bson.D{}, 0, 0)
+    if err != nil {
+        return err
+    }
+    if profiles != nil {
+        return fiber.NewError(fiber.StatusBadRequest, "Profiles binding to the node are not empty.")
+    }
     if err := models.DeleteNode(c.Params("id")); err != nil {
         return err
     }
